Fix window reference number for NSInteger usage on 64 bit
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/branches/gnustep_testplant_branch@40340 72102866-910b-0410-8b05-ffd578937521

+++ Source/x11/XGDragView.m
@@ -309,7 +309,7 @@ static	XGDragView	*sharedDragView = nil;
 
 
 - (NSWindow*) windowAcceptingDnDunder: (NSPoint)p
-			    windowRef: (int*)mouseWindowRef
+			    windowRef: (NSInteger*)mouseWindowRef
 {
   gswindow_device_t	*dwindev;
